+++ Content/Content/StartLevel.h
@@ -11,5 +11,5 @@ public:
 
 	void BeginPlay() override;
 
-	void Tick(unsigned long long deltaTime) override;
+	void Tick(double deltaTime) override;
 };
